+++ website/docusaurus.config.js
@@ -35,12 +35,8 @@ module.exports = {
           title: "Docs",
           items: [
             {
-              label: "Style Guide",
-              to   : "docs/",
-            },
-            {
-              label: "Second Doc",
-              to   : "docs/doc2/",
+              label: "Getting Started",
+              to   : "docs/getting-started",
             },
           ],
         },