+++ website/sidebars.js
@@ -1,6 +1,15 @@
 module.exports = {
   someSidebar: {
-    "Style Composer": ["getting-started", "usage", "style-rules", "pseudo-classes", "dynamic-units"],
-    Rules: ["rule-media"],
+    "Style Composer": [
+      "getting-started",
+      "usage",
+      "style-rules",
+      "pseudo-classes",
+      "dynamic-units",
+      "use-composed-style",
+      "use-composed-value"
+    ],
+    "Rules"         : [
+      "rule-media"],
   },
 };